Move model_transformer.py to transformations/semantic_manifest_transformer.py.
